% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{auth_exists}
\alias{auth_exists}
\title{Check whether authentication credentials exist}
\usage{
auth_exists()
}
\value{
Logical value indicating if base_url and access_token exist in the
  local env.
}
\description{
Checks if a BrAPI Base URL and access token exist in the current
  environment. Note that it does not specify whether the credentials are
  valid, merely that they exist.
}
