% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{filter_observations}
\alias{filter_observations}
\title{Retrieve a filtered list of observation data}
\usage{
filter_observations(
  year = NA,
  location = NA,
  exp_name = NA,
  env_name = NA,
  exp_type = NA,
  page_size = 10000,
  drop_empty_columns = FALSE,
  include_dbids = FALSE,
  case_correct = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{A year or vector of years.}

\item{location}{A location name or vector of locations.}

\item{exp_name}{An experiment name or vector of names.}

\item{env_name}{An environment name or vector of names.}

\item{exp_type}{An experiment type or vector of types.}

\item{page_size}{Page size to use for the response. Larger page sizes may
decrease total retrieval time.}

\item{drop_empty_columns}{If \code{TRUE}, all empty columns (including
metadata columns) will be dropped from the retrieved data frame.}

\item{include_dbids}{If \code{TRUE}, the alphanumeric DBIDs used by DeltaBreed
will be included in the retrieved data frame.}

\item{case_correct}{If \code{FALSE}, any categorical variables with case
mismatches to the trait variable definition (e.g. a value of "Blue" in the
observation data, when the defined trait categories are "red" and "blue")
will be masked as \code{NA} instead of being case-corrected to the
appropriate values.}

\item{verbose}{If \code{FALSE}, all non-essential messages will be suppressed.}
}
\value{
A data frame of observations using the supplied filters.
}
\description{
This function retrieves a foo filtered list of observation data
using one or more filter values provided by the user, packaging the
response into a data frame of the same format as \code{\link[=get_observations]{get_observations()}}.

This function is best used by calling \code{\link[=get_experiments]{get_experiments()}} first, in order
to see what experiments are actually present in the target DeltaBreed
instance.
}
\examples{
login_deltabreed("example", verbose = FALSE)

# the available filters (and values to filter on) correspond to columns of get_experiments()
get_experiments()

all_obs <- get_observations(verbose = FALSE)
manitoba <- filter_observations(location = "Manitoba")
table(all_obs$Location)
table(manitoba$Location)

# all filters can take vectors as arguments
almtis <- filter_observations(location = c("Alma", "Tisdale"))
table(almtis$Year, almtis$Location)

# filters can be combined as needed
ayt_mb <- filter_observations(exp_type = "AYT", location = "Manitoba")
table(ayt_mb$ExpName, ayt_mb$EnvName)
}
