% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiments}
\alias{get_experiments}
\title{Retrieve experiment/ summary}
\usage{
get_experiments(verbose = TRUE, include_dbids = FALSE)
}
\arguments{
\item{verbose}{If \code{FALSE}, all non-essential messages will be suppressed.}

\item{include_dbids}{If \code{TRUE}, the alphanumeric DBIDs used by DeltaBreed
for each experiment/environment will be included in the retrieved data
frame. Typically used for debugging or merging data from other sources.}
}
\value{
Data frame of experiment/environment metadata.
}
\description{
Retrieves a summary of all experiments and environments in a
given DeltaBreed program. This may include experiments for which no
observations have been recorded yet.
}
\examples{
login_deltabreed("example", verbose = FALSE)

expt <- get_experiments()
expt
}
