% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Retrieve all observation data}
\usage{
get_observations(
  page_size = 10000,
  drop_empty_columns = FALSE,
  include_dbids = FALSE,
  case_correct = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{page_size}{Page size to use for the response. Larger page sizes may
decrease total retrieval time.}

\item{drop_empty_columns}{If \code{TRUE}, all empty columns (including
metadata columns) will be dropped from the retrieved data frame.}

\item{include_dbids}{If \code{TRUE}, the alphanumeric DBIDs used by DeltaBreed
will be included in the retrieved data frame.}

\item{case_correct}{If \code{FALSE}, any categorical variables with case
mismatches to the trait variable definition (e.g. a value of "Blue" in the
observation data, when the defined trait categories are "red" and "blue")
will be masked as \code{NA} instead of being case-corrected to the
appropriate values.}

\item{verbose}{If \code{FALSE}, all non-essential messages will be suppressed.}
}
\value{
A data frame of all observation units and any observations
(phenotypes).
}
\description{
Retrieves all observation data from a DeltaBreed instance via
BrAPI call. This includes observation units with no observations.
converting it into a data frame that mimics the appearance of the data
tables on the Experiments & Observations tab of DeltaBreed.
}
\examples{
login_deltabreed("example", verbose = FALSE)
obs <- get_observations()

# phenotype columns will be typed according to the observation variables (trait definitions)
str(obs[,16:20])
get_variables()
}
