% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traits.R
\name{get_traits}
\alias{get_traits}
\title{Get trait data}
\usage{
get_traits(include_archived = FALSE)
}
\value{
Data frame of trait information drawn from BrAPI endpoints
}
\description{
Retrieves trait data from a DeltaBreed program via BrAPI.
}
\examples{
\dontrun{
get_traits()
}
}
