% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_variables}
\alias{get_variables}
\title{Retrieve observation variables (trait definitions) from a DeltaBreed instance.}
\usage{
get_variables(verbose = TRUE, include_archived = FALSE)
}
\arguments{
\item{verbose}{If \code{FALSE}, all non-essential messages will be suppressed.}

\item{include_archived}{If \code{TRUE}, archived (non-active) variables will be
included in the retrieved data frame.}
}
\value{
Data frame of trait definitions drawn from BrAPI /variables
endpoint.
}
\description{
Retrieves trait data from a DeltaBreed program via BrAPI,
converting it into a data frame that mimics the appearance of the Ontology
table in DeltaBreed itself.
}
\examples{
login_deltabreed("example", verbose = FALSE)

vars <- get_variables()
vars
}
