% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{check_auth}
\alias{check_auth}
\title{Validate BrAPI authentication credentials}
\usage{
check_auth()
}
\value{
No return value, called for side effects (printing status)
}
\description{
Checks if the user has credentials currently stored and
 validates them by performing a test call to the BrAPI endpoint.
 Also prints the remaining time until the access token expires, if
 applicable.
}
