% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{filter_observations}
\alias{filter_observations}
\title{Retrieve a filtered list of observation data}
\usage{
filter_observations(
  year = NA,
  location = NA,
  exp_name = NA,
  env_name = NA,
  exp_type = NA,
  page_size = 10000,
  drop_empty_columns = FALSE,
  include_dbids = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{year}{A year or vector of years.}

\item{location}{A location name or vector of locations.}

\item{exp_name}{An experiment name or vector of names.}

\item{env_name}{An environment name or vector of names.}

\item{exp_type}{An experiment type or vector of types.}

\item{page_size}{Page size to use for the response. Larger page sizes may decrease total retrieval time.}

\item{drop_empty_columns}{Whether to drop all empty columns (including metadata columns) from the returned data frame}

\item{include_dbids}{Whether to include the DbIds of the observation units, mostly useful for debugging.}

\item{verbose}{Whether to print short messages about the number of records found.}
}
\value{
A data frame of observations using the supplied filters.
}
\description{
This function retrieves a filtered list of observation data using one or more
filters provided by the user, packaging the response into a data frame of the
same format as get_observations().
}
\details{
This function is best used by calling get_experiments() first, in order to
see what experiments are actually present in the target DeltaBreed instance.
}
\examples{
login_deltabreed("example", verbose = FALSE)

# the available filters (and values to filter on) correspond to columns of get_experiments()
get_experiments()

all_obs <- get_observations(verbose = FALSE)
manitoba <- filter_observations(location = "Manitoba")
table(all_obs$Location)
table(manitoba$Location)

# all filters can take vectors as arguments
almtis <- filter_observations(location = c("Alma", "Tisdale"))
table(almtis$Year, almtis$Location)

# filters can be combined as needed
ayt_mb <- filter_observations(exp_type = "AYT", location = "Manitoba")
table(ayt_mb$ExpName, ayt_mb$EnvName)
}
