% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiments.R
\name{get_experiments}
\alias{get_experiments}
\title{Get experiment data}
\usage{
get_experiments(summarize = TRUE)
}
\value{
Data frame of experiment information drawn from BrAPI endpoints.
}
\description{
Retrieves the list of all experiments and environments in a DeltaBreed program.
}
\examples{
\dontrun{
get_experiments()
}
}
