% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germplasm.R
\name{get_germplasm}
\alias{get_germplasm}
\title{Get germplasm data}
\usage{
get_germplasm()
}
\value{
Germplasm data from the BrAPI API
}
\description{
Retrieves all germplasm data from the current DeltaBreed instance.
}
\examples{
\dontrun{
login_deltabreed()
germplasm <- get_germplasm()
}
}
