% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germplasm.R
\name{get_germplasm}
\alias{get_germplasm}
\title{Retrieve germplasm records}
\usage{
get_germplasm(page_size = 10000)
}
\arguments{
\item{page_size}{Page size to use for the response. Larger page sizes may
decrease total retrieval time.}
}
\value{
Data frame of germplasm/accession/entry information.
}
\description{
Retrieves all germplasm data from the current DeltaBreed
instance, reformatting it to match DeltaBreed layout.
}
\examples{
login_deltabreed("example", verbose = FALSE)

germplasm <- get_germplasm()
head(germplasm)
}
