% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Retrieve all observation data}
\usage{
get_observations(
  page_size = 10000,
  drop_empty_columns = FALSE,
  include_dbids = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{page_size}{Page size to use when making the request. Larger page size can make requests faster but may cause errors.}

\item{drop_empty_columns}{Whether to drop all empty columns (including metadata columns) from the returned data frame}

\item{include_dbids}{Whether to include the DbIds of the observation units, mostly useful for debugging.}

\item{verbose}{Whether to print short messages showing the number of records found.}
}
\value{
A data frame of observations (phenotypes), formatted in DeltaBreed style.
}
\description{
Retrieves all observation and sub-observation data from a DeltaBreed
instance via BrAPI call, converting it into a data frame that resembles
how the data appears on DeltaBreed itself.
}
\examples{
\dontrun{
login_deltabreed()
obs <- get_observations()
}
}
