% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Get Observation data}
\usage{
get_observations()
}
\value{
Observation data from the BrAPI API
}
\description{
Retrieves observation data from a DeltaBreed program via BrAPI.
}
\examples{
\dontrun{
get_observations()
}
}
