% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get observation variables (trait definitions) from a DeltaBreed instance.}
\usage{
get_variables(include_archived = FALSE, page_size = 1000)
}
\value{
Data frame of trait information drawn from BrAPI `/variables`
endpoint.
}
\description{
Retrieves trait data from a DeltaBreed program via BrAPI.
}
\examples{
\dontrun{
get_variables()
}
}
