% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_variables}
\alias{get_variables}
\title{Get observation variables (trait definitions) from a DeltaBreed instance.}
\usage{
get_variables(verbose = TRUE, include_archived = FALSE)
}
\arguments{
\item{verbose}{Whether to print a short message about the number of traits found.}

\item{include_archived}{Should the output include archived (non-active) traits?}
}
\value{
Data frame of trait information drawn from BrAPI `/variables`
endpoint.
}
\description{
Retrieves trait data from a DeltaBreed program via BrAPI.
}
\examples{
\dontrun{
get_variables()
}
}
