% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{login_deltabreed}
\alias{login_deltabreed}
\title{Log in to a DeltaBreed instance}
\usage{
login_deltabreed(base_url = NULL, access_token = NULL, verbose = TRUE)
}
\arguments{
\item{base_url}{The BrAPI Base URL, found on the BrAPI tab of DeltaBreed
in the BrAPI Information pane.}

\item{access_token}{A valid Access Token, retrieved from the DeltaBreed
user interface.}

\item{verbose}{Whether to print out success messages.}
}
\value{
No return value, called for side effects (storing credentials)
}
\description{
This function "logs in" to a DeltaBreed instance using the BrAPI Base
  URL and an Access Token, which are then stored. Both the URL and the token
  can be given as arguments or supplied to function prompts. The function
  performs some basic checks, including verifying that the user has internet
  access and making some test calls to the BrAPI server.
}
\details{
Access tokens are valid for 24 hours after generation. To check your
  authorization credentials at any time, use the
  check_auth() function.
}
\examples{
\dontrun{
login_deltabreed()
}
}
