% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{logout_deltabreed}
\alias{logout_deltabreed}
\title{Clear DeltaBreed authentication credentials}
\usage{
logout_deltabreed()
}
\value{
No return value, called for side effects (clearing credentials)
}
\description{
Removes stored credentials (URL and access token) from the
package environment. The access token will remain valid for as long as the
DeltaBreed instance specifies, but in order to retrieve data the URL/token
will need to be-entered with `login_deltabreed()`.
}
